# docs/openapi/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "DeletionResponse": {
        "additionalProperties": false,
        "properties": {
          "cleanup": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Cleanup",
            "type": "object"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "lifecycle_state": {
            "title": "Lifecycle State",
            "type": "string"
          },
          "visible": {
            "title": "Visible",
            "type": "boolean"
          }
        },
        "required": [
          "document_id",
          "lifecycle_state",
          "visible",
          "cleanup"
        ],
        "title": "DeletionResponse",
        "type": "object"
      },
...
      "DocumentChunkResponse": {
        "additionalProperties": false,
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "string"
          },
          "document_version_id": {
            "title": "Document Version Id",
            "type": "string"
          },
          "is_parent": {
            "default": false,
            "title": "Is Parent",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "locator": {
            "additionalProperties": true,
            "title": "Locator",
            "type": "object"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "parent_chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Chunk Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "vector_indexed": {
            "default": false,
            "title": "Vector Indexed",
            "type": "boolean"
          }
        },
        "required": [
          "chunk_id",
          "document_version_id",
          "parent_chunk_id",
          "ordinal",
          "kind",
          "token_count",
          "status",
          "text",
          "locator"
        ],
        "title": "DocumentChunkResponse",
        "type": "object"
      },
      "DocumentQualityResponse": {
        "additionalProperties": false,
        "properties": {
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "document_version_id": {
            "title": "Document Version Id",
            "type": "string"
          },
          "issue_codes": {
            "items": {
              "type": "string"
            },
            "title": "Issue Codes",
            "type": "array"
          },
          "locator_coverage": {
            "title": "Locator Coverage",
            "type": "number"
          },
          "node_count": {
            "title": "Node Count",
            "type": "integer"
          },
          "parser_revision": {
            "title": "Parser Revision",
            "type": "string"
          },
          "real_acceptance": {
            "title": "Real Acceptance",
            "type": "boolean"
          },
          "source_format": {
            "title": "Source Format",
            "type": "string"
          }
        },
        "required": [
          "document_version_id",
          "source_format",
          "parser_revision",
          "node_count",
          "locator_coverage",
          "issue_codes",
          "disposition",
          "real_acceptance"
        ],
        "title": "DocumentQualityResponse",
        "type": "object"
      },
...
      "DocumentReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          },
          "decision": {
            "enum": [
              "APPROVED",
              "NEEDS_REWORK",
              "REJECTED"
            ],
            "title": "Decision",
            "type": "string"
          },
          "security_projection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SecurityProjectionRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "decision"
        ],
        "title": "DocumentReviewRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Enterprise RAG Knowledge Base",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/document-versions/{version_id}/chunks": {
      "get": {
        "operationId": "chunks_api_v1_document_versions__version_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentChunkResponse"
                  },
                  "title": "Response Chunks Api V1 Document Versions  Version Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Chunks",
        "tags": [
          "documents",
          "retrieval"
        ]
      }
    },
    "/api/v1/document-versions/{version_id}/quality-report": {
      "get": {
        "operationId": "quality_report_api_v1_document_versions__version_id__quality_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentQualityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Quality Report",
        "tags": [
          "validation"
        ]
      }
    },
    "/api/v1/document-versions/{version_id}/review": {
      "post": {
        "operationId": "review_document_version_api_v1_document_versions__version_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Review Document Version",
        "tags": [
          "validation"
        ]
      }
    },
...
    "/api/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "lifecycle"
        ]
      },
      "get": {
        "operationId": "document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/versions": {
      "get": {
        "operationId": "versions_api_v1_documents__document_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentVersionResponse"
                  },
                  "title": "Response Versions Api V1 Documents  Document Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Versions",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/ingestion-jobs/{job_id}": {
      "get": {
        "operationId": "job_status_api_v1_ingestion_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/ingestion-jobs/{job_id}:cancel": {
      "post": {
        "operationId": "cancel_job_api_v1_ingestion_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/ingestion-jobs/{job_id}:retry": {
      "post": {
        "operationId": "retry_job_api_v1_ingestion_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}